--- src/pages/user/ChangePw.js
@@ -1,1 +1,0 @@
-//  비밀번호 변경 페이지
