--- src/pages/user/IdFind.js
@@ -1,1 +1,0 @@
-//  아이디 찾기 페이지
